pyproject.toml
[project]
name = "pulchra"
version = "v1.0.0"
description = "Protein Chain Restoration Algorithm - Full-atom model reconstruction from C-alpha traces"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Julian Koh"},
]
keywords = ["protein", "structure", "reconstruction", "modeling", "bioinformatics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "biopython>=1.79",
    "click>=8.0.0",
]

[project.scripts]
pulchra = "pulchra.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pulchra"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
